refactor : 수익률(returns) Long 타입에서 BigDecimal 타입으로 변경)

+++ src/main/java/SMU/StockMate/domain/user/dto/PortfolioResponseDTO.java
@@ -3,6 +3,7 @@ package SMU.StockMate.domain.user.dto;
 import SMU.StockMate.domain.userStock.dto.UserStockResponseDTO;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
@@ -12,6 +13,6 @@ public record PortfolioResponseDTO(
    Long totalAsset,
    Long cashBalance,
    Long stockValuation,
-   Long totalReturns,
+   BigDecimal totalReturns,
    List<UserStockResponseDTO.toPortfolioDTO> userStockList
 ) {}
